add config test, repository and bug fixes
