+++ config_test.yml.dist
@@ -1,0 +1,5 @@
+database:
+  connection: postgres://mg_telegram_test:mg_telegram_test@postgres_test:5450/mg_telegram_test?sslmode=disable
+
+http_server:
+  listen: :3001
