+++ docker-compose.yml
@@ -10,7 +10,7 @@ services:
         ports:
             - ${POSTGRES_ADDRESS:-127.0.0.1:5434}:5432
 
-    mg_manager:
+    mg_telegram:
         image: golang:1.9.3-stretch
         working_dir: /mg_telegram
         user: ${UID:-1000}:${GID:-1000}
