+++ telegram.go
@@ -6,12 +6,13 @@ import (
 	"github.com/go-telegram-bot-api/telegram-bot-api"
 )
 
-func SetMsgHandler() {
+func setMsgHandler() {
 	http.HandleFunc("/add-bot/", addBotHandler)
 	http.HandleFunc("/del-bot/", deleteBotHandler)
-	http.HandleFunc("/map-bot/", mappingBotHandler)
+	http.HandleFunc("/map-bot/", mappingHandler)
 }
 
+// Get bot info
 func GetBotInfo(token string) (*tgbotapi.BotAPI, error) {
 	bot, err := tgbotapi.NewBotAPI(token)
 	if err != nil {
@@ -20,6 +21,7 @@ func GetBotInfo(token string) (*tgbotapi.BotAPI, error) {
 	return bot, nil
 }
 
+// Get name bot
 func GetNameBot(bot *tgbotapi.BotAPI) string {
 	return bot.Self.FirstName
 }
